requirements.txt
numpy==2.3.0
Pillow==11.2.1
torch==2.4.0
